[REFACTOR] Extracted overlay view to separate file
