+++ app/src/main/java/ink/trmnl/android/ui/display/TrmnlMirrorDisplayScreen.kt
@@ -13,20 +13,11 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowForward
-import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExtendedFloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -42,9 +33,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.window.core.layout.WindowSizeClass
-import androidx.window.core.layout.WindowSizeClass.Companion.WIDTH_DP_EXPANDED_LOWER_BOUND
-import androidx.window.core.layout.WindowSizeClass.Companion.WIDTH_DP_MEDIUM_LOWER_BOUND
 import coil3.compose.AsyncImage
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.runtime.CircuitUiEvent
@@ -320,145 +308,6 @@ fun TrmnlMirrorDisplayContent(
     }
 }
 
-@Composable
-private fun OverlaySettingsView(
-    state: TrmnlMirrorDisplayScreen.State,
-    windowSizeClass: WindowSizeClass = currentWindowAdaptiveInfo().windowSizeClass,
-) {
-    // Shows larger button on tablets
-    // https://developer.android.com/develop/ui/compose/layouts/adaptive/support-different-display-sizes
-    // https://developer.android.com/develop/ui/compose/layouts/adaptive/use-window-size-classes
-    val isExpandedWidth =
-        windowSizeClass.isWidthAtLeastBreakpoint(WIDTH_DP_EXPANDED_LOWER_BOUND) ||
-            windowSizeClass.isWidthAtLeastBreakpoint(WIDTH_DP_MEDIUM_LOWER_BOUND)
-
-    // Choose text style based on window width
-    val fabTextStyle =
-        if (isExpandedWidth) {
-            MaterialTheme.typography.titleLarge
-        } else {
-            MaterialTheme.typography.bodyLarge
-        }
-
-    val infoTextStyle =
-        if (isExpandedWidth) {
-            MaterialTheme.typography.titleLarge
-        } else {
-            MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.Bold)
-        }
-
-    Card(
-        modifier =
-            Modifier
-                .padding(16.dp),
-        elevation =
-            CardDefaults.cardElevation(
-                defaultElevation = 4.dp,
-            ),
-        colors =
-            CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.surface.copy(alpha = 0.8f),
-            ),
-    ) {
-        Column(
-            modifier =
-                Modifier
-                    .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Text(
-                text = "Display Configurations",
-                style = MaterialTheme.typography.displaySmall,
-                fontWeight = FontWeight.Bold,
-                modifier = Modifier.padding(bottom = 8.dp),
-            )
-
-            Text("Display image refresh: ${state.nextImageRefreshIn}", style = infoTextStyle)
-
-            ExtendedFloatingActionButton(
-                onClick = {
-                    state.eventSink(TrmnlMirrorDisplayScreen.Event.ConfigureRequested)
-                },
-                icon = {
-                    Icon(
-                        imageVector = Icons.Default.Settings,
-                        contentDescription = null,
-                        modifier = if (isExpandedWidth) Modifier.size(32.dp) else Modifier,
-                    )
-                },
-                text = {
-                    Text(
-                        "Configure Mirror Device",
-                        style = fabTextStyle,
-                        fontWeight = FontWeight.Bold,
-                    )
-                },
-            )
-
-            ExtendedFloatingActionButton(
-                onClick = {
-                    state.eventSink(TrmnlMirrorDisplayScreen.Event.RefreshCurrentPlaylistItemRequested)
-                },
-                icon = {
-                    Icon(
-                        imageVector = Icons.Default.Refresh,
-                        contentDescription = null,
-                        modifier = if (isExpandedWidth) Modifier.size(32.dp) else Modifier,
-                    )
-                },
-                text = {
-                    Text(
-                        "Refresh Current Playlist Image",
-                        style = fabTextStyle,
-                        fontWeight = FontWeight.Bold,
-                    )
-                },
-            )
-
-            ExtendedFloatingActionButton(
-                onClick = {
-                    state.eventSink(TrmnlMirrorDisplayScreen.Event.LoadNextPlaylistItemImage)
-                },
-                icon = {
-                    Icon(
-                        imageVector = Icons.AutoMirrored.Filled.ArrowForward,
-                        contentDescription = null,
-                        modifier = if (isExpandedWidth) Modifier.size(32.dp) else Modifier,
-                    )
-                },
-                text = {
-                    Text(
-                        "Load Next Playlist Image",
-                        style = fabTextStyle,
-                        fontWeight = FontWeight.Bold,
-                    )
-                },
-            )
-
-            ExtendedFloatingActionButton(
-                onClick = {
-                    state.eventSink(TrmnlMirrorDisplayScreen.Event.ViewLogsRequested)
-                },
-                icon = {
-                    Icon(
-                        imageVector = Icons.AutoMirrored.Filled.List,
-                        contentDescription = null,
-                        modifier = if (isExpandedWidth) Modifier.size(32.dp) else Modifier,
-                    )
-                },
-                text = {
-                    Text(
-                        "View Image Refresh Logs",
-                        style = fabTextStyle,
-                        fontWeight = FontWeight.Bold,
-                    )
-                },
-            )
-        }
-    }
-}
-
 @Preview(name = "Trmnl Mirror Display Error Content Preview")
 @Composable
 fun PreviewTrmnlMirrorDisplayErrorContent() {
@@ -471,24 +320,6 @@ fun PreviewTrmnlMirrorDisplayErrorContent() {
                     nextImageRefreshIn = "5 minutes",
                     isLoading = false,
                     errorMessage = "Sample Error Message",
-                    eventSink = {},
-                ),
-        )
-    }
-}
-
-@Preview(name = "Overlay Settings Preview")
-@Composable
-fun PreviewOverlaySettingsView() {
-    Surface {
-        OverlaySettingsView(
-            state =
-                TrmnlMirrorDisplayScreen.State(
-                    imageUrl = null,
-                    overlayControlsVisible = true,
-                    nextImageRefreshIn = "5 minutes",
-                    isLoading = false,
-                    errorMessage = null,
                     eventSink = {},
                 ),
         )
